original collision implemented

+++ headers/items.hpp
@@ -2,21 +2,21 @@
 #include <raylib.h>
 
 typedef struct {
-    int radius;
+    float radius;
     Color colour;
     int ID;
 } Apple;
 
 typedef struct {
-    int w;
-    int h;
+    float w;
+    float h;
     Color colour;
     int ID;
 } Berry;
 
 extern Vector2 mousePos;
 
-extern const int tileSize;
+extern const float tileSize;
 extern const int gridWidth;
 extern const int gridHeight;
 extern int grid[][33];
